+++ src/lib/types.ts
@@ -731,6 +731,8 @@ export type Wallet = WalletData & {
   supportedLanguages: string[]
 }
 
+export type WalletRow = Wallet & { id: string }
+
 export type WalletFilter = typeof WALLETS_FILTERS_DEFAULT
 
 export interface WalletFilterData {
