+++ src/lib/utils/wallets.ts
@@ -13,7 +13,13 @@ import {
   NEW_TO_CRYPTO_FEATURES,
   NFTS_FEATURES,
 } from "../constants"
-import type { Lang, WalletData, WalletFilter } from "../types"
+import type {
+  ChainName,
+  FilterOption,
+  Lang,
+  WalletData,
+  WalletRow,
+} from "../types"
 
 export const getSupportedLocaleWallets = (locale: string) =>
   shuffle(
@@ -153,42 +159,6 @@ export const getAllWalletsLanguages = (locale: string) => {
   )
 }
 
-// Get a list of top n wallets languages
-export const getWalletsTopLanguages = (n: number, locale: string) => {
-  const compareFn = (a: string, b: string) => {
-    return getLanguageTotalCount(b) - getLanguageTotalCount(a)
-  }
-
-  return walletsData
-    .reduce(
-      (allLanguagesList, current) =>
-        // `union` lodash method merges all arrays removing duplicates
-        union(allLanguagesList, current.languages_supported),
-      [] as string[]
-    )
-    .sort(compareFn)
-    .map((languageCode) => {
-      // Get supported language name
-      const supportedLanguageName = getLanguageCodeName(languageCode, locale)
-      // Return {code, capitalized language name}
-      return {
-        code: languageCode,
-        langName: `${capitalize(
-          supportedLanguageName!
-        )} (${getLanguageTotalCount(languageCode)})`,
-      }
-    })
-    .slice(0, n)
-}
-
-// Get wallets listing count after applying filters
-export const walletsListingCount = (filters: WalletFilter) => {
-  return Object.values(filters).reduce(
-    (acc, filter) => (filter ? acc + 1 : acc),
-    0
-  )
-}
-
 export const getLanguageCountWalletsData = (locale: string) => {
   const languageCountWalletsData = getAllWalletsLanguages(locale).map(
     (language) => ({
@@ -201,15 +171,52 @@ export const getLanguageCountWalletsData = (locale: string) => {
   return languageCountWalletsData
 }
 
-export const getFilteredWalletsCount = (
-  wallets: WalletData[],
-  filters: WalletFilter
-) => {
-  return wallets.filter((wallet) => {
-    const activeFilters = Object.entries(filters).filter(
-      ([, value]) => value === true
-    )
+function getActiveFilterKeys(filters: FilterOption[]): string[] {
+  const keys: string[] = []
+  filters.forEach((filter) => {
+    filter.items.forEach((item) => {
+      if (item.inputState === true && item.options.length === 0) {
+        keys.push(item.filterKey)
+      }
+      if (item.options?.length > 0) {
+        item.options.forEach((option) => {
+          if (option.inputState === true) {
+            keys.push(option.filterKey)
+          }
+        })
+      }
+    })
+  })
+  return keys
+}
 
-    return activeFilters.every(([feature]) => wallet[feature] === true)
-  }).length
+export const filterFn = (data: WalletRow[], filters: FilterOption[]) => {
+  let selectedLanguage: string = ""
+  let selectedLayer2: ChainName[] = []
+
+  const activeFilterKeys = getActiveFilterKeys(filters)
+
+  filters.forEach((filter) => {
+    filter.items.forEach((item) => {
+      if (item.filterKey === "languages") {
+        selectedLanguage = item.inputState as string
+      } else if (item.filterKey === "layer_2_support") {
+        selectedLayer2 = (item.inputState as ChainName[]) || []
+      }
+    })
+  })
+
+  return data
+    .filter((item) => {
+      return item.languages_supported.includes(selectedLanguage as Lang)
+    })
+    .filter((item) => {
+      return (
+        selectedLayer2.length === 0 ||
+        selectedLayer2.every((chain) => item.supported_chains.includes(chain))
+      )
+    })
+    .filter((item) => {
+      return activeFilterKeys.every((key) => item[key])
+    })
 }
